analytics-service/requirements.txt
PyGithub
c8y_api
flask
python-dotenv